[DebuggerV2] Refactor container tests for clearer directory organization
