+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/timeline/BUILD
@@ -31,13 +31,25 @@ ng_module(
 )
 
 tf_ts_library(
-    name = "timeline_test",
+    name = "timeline_container_test_lib",
     testonly = True,
     srcs = [
-        "timeline_test.ts",
+        "timeline_container_test.ts",
     ],
+    tsconfig = "//:tsconfig-test",
     deps = [
         ":timeline",
+        "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin:debugger_v2",
+        "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/actions",
+        "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/store",
+        "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/store:types",
+        "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/testing",
+        "//tensorboard/webapp/angular:expect_angular_core_testing",
+        "@npm//@angular/common",
+        "@npm//@angular/compiler",
+        "@npm//@angular/core",
+        "@npm//@angular/platform-browser",
+        "@npm//@ngrx/store",
         "@npm//@types/jasmine",
     ],
 )
